+++ src/main/java/org/accompany/backend/domain/deceasedProfile/dto/response/DeceasedSurveyStatusRes.java
@@ -1,6 +1,6 @@
 package org.accompany.backend.domain.deceasedProfile.dto.response;
 
-import org.accompany.backend.domain.user.entity.SurveyStatus;
+import org.accompany.backend.domain.deceasedProfile.entity.SurveyStatus;
 
 public record DeceasedSurveyStatusRes(
         SurveyStatus surveyStatus
